+++ src/harlequin_duckdb/interactions.py
@@ -22,3 +22,23 @@ def execute_use_statement(
         raise
     else:
         driver.notify(f"Editor context switched to {item.label}")
+
+
+def execute_drop_database_statement(
+    item: "InteractiveCatalogItem",
+    connection: "DuckDbConnection",
+    driver: "HarlequinDriver",
+) -> None:
+    def _drop_database() -> None:
+        try:
+            connection.execute(f"drop database {item.qualified_identifier}")
+        except HarlequinQueryError:
+            driver.notify("Could not drop database", severity="error")
+            raise
+        else:
+            driver.notify(f"Dropped database {item.label}")
+
+    if item.children or item.fetch_children(connection=connection):
+        driver.confirm_and_execute(callback=_drop_database)
+    else:
+        _drop_database()